\name{setProperties-methods}
\docType{methods}
\alias{setProperties}
\alias{setProperties-methods}
\alias{setProperties,GenesRanking-method}
\alias{setProperties.GenesRanking}
\title{Set properties}
\description{ Allows setting or modifiying the GenesRanking properties.
}
\section{Methods}{
\describe{
\code{setProperties(object, geneLabels=NULL, discriminantPower=NULL, }\cr
\code{meanDif=NULL, isRedundant=NULL, gERankMean=NULL)}
}}
\seealso{
Main package function and classifier training:
\code{\link[geNetClassifier:geNetClassifier]{geNetClassifier}} \cr
This method's class (\code{\link[geNetClassifier:GenesRanking-class]{GenesRanking}}) help page.
}

\keyword{methods}












