\name{GenesRanking-class}
\Rdversion{1.1}
\docType{class}
\alias{GenesRanking}
\alias{GenesRanking-class}
\alias{extractGenes,GenesRanking-method}
%\alias{getTopRanking,GenesRanking-method}
%\alias{gClasses,GenesRanking-method}
%\alias{genesDetails,GenesRanking-method}
%\alias{getRanking,GenesRanking-method}
\alias{initialize,GenesRanking-method}
%\alias{nGenes,GenesRanking-method}
%\alias{numSignificantGenes,GenesRanking-method}
%\alias{overview,GenesRanking-method}
%\alias{plot,GenesRanking,missing-method}
%\alias{setProperties,GenesRanking-method}
\alias{show,GenesRanking-method}

\title{Class "GenesRanking"}
\description{
Contains the genes ranking generated by \code{\link[geNetClassifier:geNetClassifier]{geNetClassifier}} and the genes details.\cr
(Slots: \code{@classificationGenes} and \code{@genesRanking} )
}
\section{Methods}{
  \describe{
    \item{genesDetails}{\code{signature(object = "GenesRanking")}: Returns data.frames with information about the genes. }
    \item{getRanking}{\code{signature(object = "GenesRanking")}: Returns a matrix containing the ranked genes. }
    \item{getTopRanking}{\code{signature(object = "GenesRanking", numGenesClass)}: Returns a new GenesRanking object containing only the top genes of each class. }
    \item{gClasses}{\code{signature(object = "GenesRanking")}: Returns the classes for which the genes are ranked.}
    \item{nGenes}{\code{signature(object = "GenesRanking")}: Returns the number of available ranked genes per class. }
    \item{numSignificantGenes}{\code{signature(object = "GenesRanking")}: Returns the number of significant genes per class (genes over the given posterior probability threshold).}
    \item{plot}{\code{signature(x = "GenesRanking", y = "missing")}: Plots the genes' posterior probability. Wrapper of \code{\link[geNetClassifier:calculateGenesRanking]{calculateGenesRanking}}.}
 }
}
\author{
Bioinformatics and Functional Genomics Group. Centro de Investigacion del Cancer (CIC-IBMCC, USAL-CSIC). Salamanca. Spain
}
\seealso{
Main package function and classifier training:
\code{\link[geNetClassifier:geNetClassifier]{geNetClassifier}}\cr
Plot the ranking genes's posterior probability: \code{\link[geNetClassifier:plot.GenesRanking]{plot.GenesRanking}}
}
\examples{
######
# Load data and obtain a genesRanking
######

# Load an expressionSet:
library(leukemiasEset)
data(leukemiasEset)

# Select the train samples: 
trainSamples<- c(1:10, 13:22, 25:34, 37:46, 49:58) 
# summary(leukemiasEset$LeukemiaType[trainSamples])

# Train a classifier or load a trained one:
# leukemiasClassifier <- geNetClassifier(leukemiasEset[,trainSamples], 
#     sampleLabels="LeukemiaType", plotsName="leukemiasClassifier") 
data(leukemiasClassifier) # Sample trained classifier

# Or directly obtain a genesRanking with calculateGenesRanking()
\dontrun{genesRanking <- calculateGenesRanking(leukemiasEset[,trainSamples], 
    sampleLabels="LeukemiaType", returnRanking="full")}

######
# Exploring the rankings
######
genesRanking <- leukemiasClassifier@genesRanking
# Number of available genes in the ranking:
numGenes(genesRanking)
# Number of significant genes (genes with posterior probability over the threshold. 
# Default: lpThreshold=0.75):
numSignificantGenes(genesRanking)
# Top 10 genes of CML:
genesDetails(genesRanking)$CML[1:10,]
# To get a sub ranking with the top 10 genes:
getTopRanking(genesRanking, 10)
# Genes details of the top 10 genes:
genesDetails(getTopRanking(genesRanking, 10))	

###
# Genes used for training the classifier:
###
numGenes(leukemiasClassifier@classificationGenes)
leukemiasClassifier@classificationGenes
#genesDetails(leukemiasClassifier@classificationGenes)  # List by classes
genesDetails(leukemiasClassifier@classificationGenes)$AML # Show a class genes
# If your R console wraps the table rows, try widening your display width: 
# options(width=200)
}
\keyword{classes}
